maj 18 juin 2025

+++ database/migrations/2025_06_16_191042_add_role_to_users_table.php
@@ -12,14 +12,18 @@ return new class extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('role')->default('utilisateur'); // Valeur par défaut
+            if (!Schema::hasColumn('users', 'role')) {
+                $table->string('role')->default('utilisateur'); // Ajoute la colonne uniquement si elle n'existe pas
+            }
         });
     }
     
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role');
+            if (Schema::hasColumn('users', 'role')) {
+                $table->dropColumn('role'); // Supprime la colonne uniquement si elle existe
+            }
         });
     }
 };
